pyproject.toml
[tool.poetry]
name = "ai-ebook-site"
version = "0.1.0"
description = ""
authors = ["pedro.grilo <pedro.sgrilo@decsis.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.7"
langchain-community = "^0.0.38"
langchain = "^0.1.20"
slowapi = "^0.1.9"
docxtpl = "^0.17.0"
pypdf = "^4.2.0"
docx2pdf = "^0.1.8"
pypandoc = "^1.13"
pypandoc-binary = "^1.13"
boto3 = "^1.34.109"
aspose-words = "^24.5.0"
resend = "^1.1.0"
stripe = "^9.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
